requirements.txt
face_recognition==1.3.0
Pillow==9.3.0
scikit_learn==1.2.0
tqdm==4.64.1
